+++ packages/firebase_crashlytics/test/firebase_crashlytics_test.dart
@@ -44,12 +44,13 @@ void main() {
         ],
         context: ErrorDescription('foo context'),
       );
-      crashlytics.enableInDevMode = true;
-      crashlytics.log('foo');
-      crashlytics.setBool('testBool', true);
-      crashlytics.setInt('testInt', 42);
-      crashlytics.setDouble('testDouble', 42.0);
-      crashlytics.setString('testString', 'bar');
+      crashlytics
+        ..enableInDevMode = true
+        ..log('foo')
+        ..setCustomKey('testBool', true)
+        ..setCustomKey('testInt', 42)
+        ..setCustomKey('testDouble', 42.0)
+        ..setCustomKey('testString', 'bar');
       await crashlytics.recordFlutterError(details);
       expect(log[0].method, 'Crashlytics#onError');
       expect(log[0].arguments['exception'], 'foo exception');
@@ -57,18 +58,10 @@ void main() {
       expect(log[0].arguments['information'], 'test message\nsecond message');
       expect(log[0].arguments['logs'], isNotEmpty);
       expect(log[0].arguments['logs'], contains('foo'));
-      expect(log[0].arguments['keys'][0]['key'], 'testBool');
-      expect(log[0].arguments['keys'][0]['value'], isTrue);
-      expect(log[0].arguments['keys'][0]['type'], 'boolean');
-      expect(log[0].arguments['keys'][1]['key'], 'testInt');
-      expect(log[0].arguments['keys'][1]['value'], 42);
-      expect(log[0].arguments['keys'][1]['type'], 'int');
-      expect(log[0].arguments['keys'][2]['key'], 'testDouble');
-      expect(log[0].arguments['keys'][2]['value'], 42.0);
-      expect(log[0].arguments['keys'][2]['type'], 'double');
-      expect(log[0].arguments['keys'][3]['key'], 'testString');
-      expect(log[0].arguments['keys'][3]['value'], 'bar');
-      expect(log[0].arguments['keys'][3]['type'], 'string');
+      expect(log[0].arguments['keys']['testBool'], 'true');
+      expect(log[0].arguments['keys']['testInt'], '42');
+      expect(log[0].arguments['keys']['testDouble'], '42.0');
+      expect(log[0].arguments['keys']['testString'], 'bar');
     });
 
     test('recordError', () async {
@@ -81,18 +74,10 @@ void main() {
       expect(log[0].arguments['context'], "context");
       expect(log[0].arguments['logs'], isNotEmpty);
       expect(log[0].arguments['logs'], contains('foo'));
-      expect(log[0].arguments['keys'][0]['key'], 'testBool');
-      expect(log[0].arguments['keys'][0]['value'], isTrue);
-      expect(log[0].arguments['keys'][0]['type'], 'boolean');
-      expect(log[0].arguments['keys'][1]['key'], 'testInt');
-      expect(log[0].arguments['keys'][1]['value'], 42);
-      expect(log[0].arguments['keys'][1]['type'], 'int');
-      expect(log[0].arguments['keys'][2]['key'], 'testDouble');
-      expect(log[0].arguments['keys'][2]['value'], 42.0);
-      expect(log[0].arguments['keys'][2]['type'], 'double');
-      expect(log[0].arguments['keys'][3]['key'], 'testString');
-      expect(log[0].arguments['keys'][3]['value'], 'bar');
-      expect(log[0].arguments['keys'][3]['type'], 'string');
+      expect(log[0].arguments['keys']['testBool'], 'true');
+      expect(log[0].arguments['keys']['testInt'], '42');
+      expect(log[0].arguments['keys']['testDouble'], '42.0');
+      expect(log[0].arguments['keys']['testString'], 'bar');
     });
 
     test('crash', () {