+++ packages/firebase_crashlytics/CHANGELOG.md
@@ -8,6 +8,7 @@
   * `setUserName`
   * `getVersion`
   * `isDebuggable`
+* **Breaking**: the methods `setInt`, `setDouble`, `setString` and `setBool` have been replaced by `setCustomKey`.
 * Fixes a bug that prevented keys from being set on iOs devices.
 
 ## 0.1.3+3